move shared folder to its own library
