+++ library/radius-examples/components/inline-link/inline-link.styles.ts
@@ -1,5 +1,5 @@
 import { radiusTokens } from '@rangle/radius-foundations/generated/design-tokens.constants';
-import { css } from '../../../../shared/utils';
+import { css } from '@rangle/radius-shared/utils';
 import { RadiusInlineLinkExtendedProps } from './inline-link.types';
 
 export type StyleProps = Pick<RadiusInlineLinkExtendedProps, 'disabled'>;
