+++ library/radius-examples/components/link-button/link-button.styles.ts
@@ -1,5 +1,5 @@
 import { radiusTokens } from '@rangle/radius-foundations/generated/design-tokens.constants';
-import { css } from '../../../../shared/utils';
+import { css } from '@rangle/radius-shared/utils';
 import { RadiusLinkButtonExtendedProps } from './link-button.types';
 
 export type StyleProps = Pick<RadiusLinkButtonExtendedProps, 'disabled'>;
