+++ library/radius-examples/components/link-icon/link-icon.styles.ts
@@ -1,5 +1,5 @@
 import { radiusTokens } from '@rangle/radius-foundations/generated/design-tokens.constants';
-import { css } from '../../../../shared/utils';
+import { css } from '@rangle/radius-shared/utils';
 import { RadiusLinkIconExtendedProps } from './link-icon.types';
 
 export type StyleProps = Pick<RadiusLinkIconExtendedProps, 'disabled'>;
