+++ library/shared/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "@rangle/radius-shared",
+  "version": "0.0.1",
+  "type": "commonjs",
+  "files": [
+    "constants/",
+    "utils/*"
+  ]
+}
