+++ library/shared/utils/flatten.utils.ts
